Organizing code & folders. Moving "Loader" to "common" folder, simplyfying code 7 organizing files in "movielist" folder
