+++ src/features/movieList/components/ImageDefault/styled.js
